fix(msc): do not truncate the USB sector count to 16 bits
GET_SECTOR_COUNT cast tuh_msc_get_block_count() to (WORD) before storing
the result in a DWORD, capping the reported capacity at 65535 sectors
(~32 MiB). Remove the cast so the full 32-bit block count reaches FatFs.

+++ src/mia/usb/msc.c
@@ -160,7 +160,7 @@ DRESULT disk_ioctl(BYTE pdrv, BYTE cmd, void *buff)
     case CTRL_SYNC:
         return RES_OK;
     case GET_SECTOR_COUNT:
-        *((DWORD *)buff) = (WORD)tuh_msc_get_block_count(dev_addr, lun);
+        *((DWORD *)buff) = tuh_msc_get_block_count(dev_addr, lun);
         return RES_OK;
     case GET_SECTOR_SIZE:
         *((WORD *)buff) = (WORD)tuh_msc_get_block_size(dev_addr, lun);
